requirements.txt
python-dotenv==1.0.1
python-telegram-bot==21.6
openai==1.51.2
gspread==6.1.4
google-auth==2.35.0
httpx==0.27.2
fastapi==0.115.6
uvicorn==0.32.1
